api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
openai>=1.50.0
pillow==10.2.0
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
supabase>=2.0.0
